+++ packages/micropip/src/micropip/_compat_not_in_pyodide.py
@@ -30,6 +30,9 @@ class pyodide_js_:
         raise RuntimeError(f"Attempted to access property '{attr}' on pyodide_js dummy")
 
 
+REPODATA_INFO: dict[str, str] = {}
+
+
 def loadPackage(packages: str | list[str]) -> None:
     pass
 
@@ -38,6 +41,7 @@ __all__ = [
     "loadDynlib",
     "fetch_bytes",
     "fetch_string",
+    "REPODATA_INFO",
     "REPODATA_PACKAGES",
     "loadedPackages",
     "loadPackage",