+++ pyodide-build/pyodide_build/buildpkg.py
@@ -449,6 +449,7 @@ def compile(
             ldflags=build_metadata["ldflags"],
             target_install_dir=target_install_dir,
             replace_libs=replace_libs,
+            exports=build_metadata.get("exports", "pyinit"),
         )